requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.4
django-celery==3.1.17
django-environ==0.4.0
djangorestframework==3.3.3
kombu==3.0.35
psycopg2==2.6.1
pytz==2016.3
six==1.10.0
sqlparse==0.1.19
